# Generalize optimized categorical grouping method to multiple keys