The compiler is still able to generate efficient code specialized on the number of keys. Also require `groups !== nothing` for this method since we never call it without groups and it cannot compute the actual number of nonempty groups without doing all the work anyway.

```julia
using DataFrames, BenchmarkTools
df1 = DataFrame(key1=categorical(string.(rand(1:10, 500_000))),
                key2=categorical(string.(rand(1:10, 500_000))));
df2 = DataFrame(key1=categorical(string.(rand(1:1000, 500_000))),
                key2=categorical(string.(rand(1:1000, 500_000))));

# Master
julia> @btime groupby(df1, :key1);
  2.807 ms (66 allocations: 7.63 MiB)

julia> @btime groupby(df2, :key1);
  4.727 ms (68 allocations: 7.66 MiB)

julia> @btime groupby(df1, [:key1, :key2]);
  21.069 ms (499949 allocations: 23.08 MiB)

julia> @btime groupby(df2, [:key1, :key2]);
  57.993 ms (106682 allocations: 23.08 MiB)

# This PR
julia> @btime groupby(df1, :key1);
  2.324 ms (50 allocations: 7.63 MiB)

julia> @btime groupby(df2, :key1);
  4.241 ms (50 allocations: 7.66 MiB)

julia> @btime groupby(df1, [:key1, :key2]);
  3.381 ms (52 allocations: 7.63 MiB)

julia> @btime groupby(df2, [:key1, :key2]);
  38.733 ms (58 allocations: 22.23 MiB)
```